Fix compiler warning: implicit conversion loses integer precision
src/gtk_combofix.c:1390:14: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    length = strlen (string);
           ~ ^~~~~~~~~~~~~~~
src/gtk_combofix.c:1418:35: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
            model_string_length = strlen ( model_string );
                                ~ ^~~~~~~~~~~~~~~~~~~~~~~
src/gtk_combofix.c:2507:14: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
    length = strlen ( string );
           ~ ^~~~~~~~~~~~~~~~~
src/gtk_combofix.c:2516:31: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
        model_string_length = strlen ( model_string );
                            ~ ^~~~~~~~~~~~~~~~~~~~~~~

+++ src/gtk_combofix.c
@@ -1378,7 +1378,7 @@ static gchar *gtk_combofix_update_visible_rows ( GtkComboFix *combofix,
     GtkTreePath *path;
     GtkTreeIter iter;
     gint path_ok;
-    gint length;
+    gsize length;
     gboolean separator = FALSE;
     GtkComboFixPrivate *priv = combofix -> priv;
 
@@ -1401,7 +1401,7 @@ static gchar *gtk_combofix_update_visible_rows ( GtkComboFix *combofix,
         gchar *model_string;
         gint showed;
         gint show_row = 0;
-        gint model_string_length;
+        gsize model_string_length;
 
         gtk_tree_model_get ( model,
                         &iter,
@@ -1427,7 +1427,7 @@ static gchar *gtk_combofix_update_visible_rows ( GtkComboFix *combofix,
 			{
 				gchar *tmp_str1;
 				gchar *tmp_str2;
-				gint len;
+				gsize len;
 
 				len = MIN ( length, model_string_length );
 				tmp_str1 = g_utf8_casefold ( model_string, len );
@@ -2499,10 +2499,10 @@ gchar *gtk_combofix_get_first_string_from_sort_model ( GtkTreeModel *sort_model,
 {
     GtkTreeIter iter;
     GtkTreeIter child_iter;
-    gint length;
+    gsize length;
     gboolean separator = FALSE;
     gchar *model_string;
-    gint model_string_length;
+    gsize model_string_length;
 
     length = strlen ( string );
